test-app/pyproject.toml
[tool.poetry]
name = "test-app"
version = "0.1.0"
description = "Test application"
authors = ["Test Author <test@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
test-app-cli = "app.cli:main"
dev = "run:main"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-cors = "^5.0.0"

flask-sqlalchemy = "^3.1.1"
sqlalchemy = "^2.0.35"
psycopg = {extras = ["binary"], version = "^3.2.3"}
alembic = "^1.14.0"

pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
spectree = "^1.4.2"
dependency-injector = "^4.45.0"
prometheus-client = "^0.21.0"
waitress = "^3.0.2"
PasteScript = "^3.6.0"
python-dotenv = "^1.0.1"
Pillow = "^11.1.0"
httpx = "^0.28.0"
requests = "^2.32.0"

pyjwt = {extras = ["crypto"], version = "^2.10.1"}
cryptography = "^44.0.0"
itsdangerous = "^2.2.0"


boto3 = "^1.35.76"
boto3-stubs = {extras = ["s3"], version = "^1.35.76"}
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
ruff = "^0.8.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]
